Fix to #6040 - Queries projecting entities incorrectly try to inline parts of their projection even when client side projection of those elements is required
Problem was that during Projection translation we would try to inline single element subqueries into the parent query. However, in case the query source was mapped to Entity as opposed to ValueBuffer we were not modifying the selector expression afterwards.
This resulted in the same subquery being processed again, later in the pipeline, producing redundant and confusing query plans (also the inlined queries were never used in the final result)

Fix is to only inline queries for which we know how to modify the selector (i.e. single element that can be extracted from a ValueBuffer)