Fix #6122 - LoadAsync does not work
Problem: Our async Select operator was too lazy. The projector delegate was only run when the Current property was accessed, but the code in LoadAsync spins the enumerator without accessing the Current property.
Solution: Change our async Select operator to invoke the projector delegate on MoveNext. This is preferred over accessing Current during LoadAsync because it matches L2O behavior and will prevent future instances of this bug.