Fix to #4858 - Select after complex GroupJoin lead to unpredictable result
Problem exists for complex query with groupjoin, where the inner element is a subquery, then we apply collection operator on the inner collection, but project only the outer element like so:

context.LevelOne
  .GroupJoin(
    context.LevelTwo.Where(l2 => l2.Name != "L2 01"),
    l1 => l1.Id,
    l2 => l2.Level1_Optional_Id,
    (l1, l2s) => new { l1, l2s })
  .Where(r => r.l2s.Any())
  Select(r => r.l1);

What happens is that we look at which query sources we need to materialize (based on final projection), deduce that we don't need to materialize the elements of the inner collection, so it's being represented as value buffer. Problem is that currently value buffer can't represent a null-value element, so Any() always returns true (even for empty collection), which leads to incorrect results.

Fix is to force materialization of inner collection elements as entities. We already did that, but missed the case where the inner collection was a subquery.