Handle fixup of overlapping composite keys which get propagated store generated values
Issue #6212

This happens only in cases where the FK property is getting a store generated value by propagation, and this FK property is shared between two relationships, and if the metadata is built in a way that the propagation happens in a certain order. The fix is to to also use the pre-store generated value when looking up the new principal since this may not have been updated yet in the dependent map.

Also added more coverage for store-generated composite keys in general, and added some extension methods to make it easier to dump out a human-readable form of the model.

Also, see #6292, uncovered will testing.