Make operator generation more flexible
Current operator generation in queries is inflexible - the current logic
selects an operator based only on the expression's NodeType. Cases where a
different operator needs to be generated based on others factors (e.g.
the expression's Type) are difficult to integrate.

Made it more easy for providers to select operators based on any factor on
the expression.