Correct navigation index calculation
Issues #6570

This issue is caused by incorrectly calculating the indexes for navigations. Navigations started counting at the end of the properties, but I'm not sure why that was since all uses appear to be independent. This lack of overlap was then "corrected for" when using the index for the IsLoaded flag, but this correction was done incorrectly when derived types were involved. The fix is to just make navigation indexes count independently from property indexes. Note, however, that relationship indexes are still combined.