Implement complex type definitions
This PR contains the core metadata for complex type definitions. It does not include any code for the usage of these definitions in an entity type.

(Not intending to merge to this branch, but still waiting on #6543 to be reviewed so putting this out for review here to keep me unblocked.)