Fix to #6620 - GroupBy multiple keys throws exception in 1.0.1
Problem was with Tracking queries involving GroupBy and OrderBy/Skip/Take/Distinct. When tracking results of group by we were trying to extract Grouping element from the result however we always assumed that the Grouping result operator would be on the main QueryModel. When Orderby/Take/Skip etc are used in conjunction with GroupBy, the GroupBy result operator is actually in the subquery. Fix is to properly extract Group Element information from the subquery when needed.

CR: Andrew