Fix to #5544 - Query :: force client evaluation on First() inside a query
We used to try server-evaluate calls to First() in a subquery, however we are unable to properly mimic the exception that could potentially be thrown from First() method. Fix is to always force client evaluation when First() is used in a subquery.
If one wants to avoid client evaluation, FirstOrDefault() should be used.