Metadata cleanup following hierarchy/base class changes
These are changes that have been made on the ComplexTypes branch for things that I found following the change to the metadata types hierarchy and the move of some members to IPropertyBase. Making these changes to dev since they are strictly related to complex type support and it is better to change them sooner rather than later.