Fix breaking change to RelationalDatabaseCreator constructor
Issue #6622

The approach we discussed didn't pan out exactly because IRelationalDatabaseCreator can also be obtained as the non-relational service IDatabaseCreator. The delegate for this is defined in the core package and so cannot directly reference relational types. Therefore, I created an abstract interface that any provider service can implement which then allows the service itself to request the additional services it needs.

Tested by making SQLite and SQL Server providers use the obsolete constructor and running all tests.

RelationalQueryContextFactory was not changed--see issue #6713.

We should review other constructors for changes--see issue #6714