Fix to 6702 - Query: Null semantics issue after improved nav. translation
Problem was that we were not adding null semantics compensation for some cases that were enabled after LOJ translation to SQL.
Specifically, when we decide whether to apply null semantics we look at the expression nullability, and for Column/Alias expression we also look at the IProperty to see if it's nullable (e.g. in case of a non-nullable string).
LOJ translation enabled scenarios where the IProperty is non-nullable, but the actual value can be null, because it comes from an optional navigation - we should be applying null semantics for those cases.

Fix is to add extra metadata to ColumnExpression telling us when the column can be null, even if the IProperty says otherwise. The field is being set during ConditionalAccessExpression translation.

Also fixed minor bugs found along the way:
- we were incorrectly applying one optimization to a query comparing negated nullable column to constant true/false
- we were not compensating for type difference between elements of ConditionalExpression during nav reweite