Fix to #6451 - Query :: Include doesn't work with SelectMany
Problem was that we were not able to match query source we were trying to include, if it came from the SelectMany since it has different structure than just regular filter/projection etc.
Fix is to compensate for the SelectMany case when constructing IncludeSpecification by using the correct query source.