Fix SQL Server value generation from model issue
When deciding which value generation to use, the SQL Server provider falls
back to the model annotation. However, it did not properly check that the
property's CLR type is compatible with the model-determined strategy. As a
result, defining a ValueGeneratedOnAdd decimal on a model with strategy
HiLo triggered a failure.