Fix to #6896 - Query: Rewrite Principal key access to FK access only if FK entity is not projected out
Problem was that during nav rewrite we sometimes can simplify navigation like: order.Customer.Id into order.CustomerId.
This however is problematic for some complex queries like so:

var query = context.LevelThree
    .Select(l3 => l3.OneToOne_Required_FK_Inverse)
    .OrderBy(l3 => l3.Id)
    .Take(10)
    .Select(l2 => l2.OneToOne_Required_FK_Inverse.Name);

In this case we try to lift the orderby to the outer query. If the orderby gets simplified the query becomes invalid, because now we try to order by Level3.FK_ID, but what we project out from the subquery is Level2.
Fix is to not perform this optimization for orderby if we project the entity on which property we try to order.