Fix to #6861 - Query with projection and navigation properties adds extra sort columns in ORDER BY for foreign key columns
When creating LEFT OUTER JOIN we add ORDER BY clause(s) based on the key of the outer entity. This is because out GroupJoin method is streaming and we need to have results in order to create the correct groups in a performant way.
However we can there are some optimizations we can do:
1. we don't need ORDER BY if the join is from dependent to principal because there will always be at most one customer associated with a given order.
e.g. orders.Select(o => o.Customer.Name)
which translates to:
from o in Orders
join c in Customers on o.CustomerId equal c.Id into grouping
from c in grouping.DefaultIfEmpty()
select c.Name

In case of composite keys we need to check that entire key is being used as a join condition. Moreover, if there are additional elements in the join condition (on top of the FK->PK) we don't need to add order by either, because the grouping will be even more restrictive.

2. in case when we fully translate LOJ into SQL (which is now very often the case) we also don't need the extra ORDER BYs because we no longer perform streamed grouping on the client.