Fix to #7061 - the binary operator Equal is not defined for the types 'System.Nullable`1[System.Boolean]' and 'System.Boolean'
Problem was that when we were fixing search conditions on a query, we did not take into account differences in nullability that could be created by those transformations.

e.g.:
myEntity.MyNullableBool (result type is bool?)

would get converted to:
myEntity.MyNullableBool == (bool?)true (result type bool)

If that in turn was a part of a bigger expression, e.g. AndAlso with another bool? expression, that did not have to be converted, then we would produce a type discrepancy which was resulting in an error.

Fix is to adjust types of the Left and Right expressions that we process before reconstructing a binary expression node.