Proof-of-concept for re-writing ServiceCollection.
This is an idea @anpete had to re-write the service collection in the equivalent of AddEntityFramework so that in cases where we need to inject a new service we can do so. This is working, but the problem is that it only works if the service has been registered as a concrete type by the provider. If it is registered as an object or delegete, then we don't have a type that we can use to re-register the original registration before overwriting the current one with our delegate. Any ideas?