Implement service-injection mechanism for patch/point releases
Issue #7465

This change introduces a new class ServiceCollectionMap that is used by providers and by our service-adding methods. The new class builds a map from service type to indexes of the ServiceDescriptors in the list for the given type. This allows all the TryAdd calls that we make to work without scanning the list every time.

In addition, this type has a method for re-writing services to do property injection when we need to add a service dependency without breaking a constructor in a patch or point release. This is used by making a call at the end of our service registrations like so:

```C#
   ...
   .TryAddScoped<IResultOperatorHandler, ResultOperatorHandler>()
   .DoPatchInjection<IValueGeneratorSelector>();
```

Also, ReplaceService code has been updated to ensure that service-injection is also run on any replaced services, since these services may also inherit from our base classes.