Add parameter objects for dependencies of public services
Issue #7465

This change introduces parameter objects for the service dependencies of all public services. This allows new dependencies to be added without breaking derived classes that call the base class constructor. For factories, these dependency objects are passed down into the services created by the factory. This means that we also don't break by adding parameters to the constructors called by the factory. This is the main reason for using this approach rather than just the service rewriting approach.

The pattern is not used for internal services since no code should be inheriting from these classes. However, the service re-writer code is still in place, which means that we have some flexibility to add dependencies to internal services even in patch releases where we are being very strict about binary compatibility.

Some work remains still to do:
- Add With methods to all ...Dependencies objects to allow for service replacement
- Add sugar for registration
- Re-visit service scopes