Fix to #7647 - Query/Cleanup : remove redundant compilations of subquery models when processing MainFromClause, AdditionalFromClause and Join/GroupJoin clauses
This is small optimization for some queries containing subqueries. Before we used to compile them twice, throwing away the first one. This makes it harder to debug code, but also prevents translating GroupJoin queries inside subqueries.

Fix is to reverse the order in which we process those subqueries, this way we only process the subquery twice if the first pass fails (i.e. we can't translate it fully), rather than always.