Query: Refactor RequiresMaterializationExpressionVisitor
Improved the algorithm of RequiresMaterializationExpressionVisitor to
provide better detection of which query sources need materialized.
Currently the QueryCompilationContext forces all group joins to be
materialized, but these changes produce the correct results for when the
time comes to remove the forcing. These changes also prepare for the
possibility of providers translating the GroupBy result operator.