Remove special handling for concat operator
A special string concat operator was added at some point to support
PostgreSQL (which uses || to concat strings). This is an
overridable property that is by default +.

After the introduction of a more generic GenerateOperator(), this
no longer makes - a provider that needs a special concat operator
(or any other) can simply override GenerateOperator() and do that
there without any special support from Relational.

Removing to simplify/cleanup DefaultQuerySqlGenerator.