Fix to #7714 - Invalid query produced when projecting DTO containing subquery that returns a single element
Problem was that in the fix for #7033 we force client evaluation on queries that try to bind to parents, when the parent has a client projection.
However, for the case when single element is returned (e.g. cusotmers.Select(c => new DTO { Count = c.Orders.Count() })) we don't need to do client evaluation.
We actually try to merge the subquery into the parent query, but this causes a problem because earlier in the pipeline we assumed that client eval is needed (since parent QM has client projection)

Fix is to be consistent and never try to lift subquery if the parent QM has a client projection. This causes more scenarios to produce N+1 queries, but the proper fix would require a breaking change:
we need additional information passed to the SqlTranslatingExpressionVisitor to know that the subquery that is being visited returns a single value.