Fix to #6611 - Hundred of queries being ran. When calling OrderBy, then Include then Where to a one to many relationship
Problem was that during initial parsing of query model, Include, AsTracking, AsNoTracking force subqueries to be created (since they are result operators). Those operators are then removed, but the subqueries persist, making it much harder to translate queries efficiently. We try to prune "empty" subqueries, but sometimes it is safe to lift subqueries that contain body clauses.

Fix is to recognize those patterns and perform the lift during query model optimizations. We do it for cases when the subquery is coming from the MainFromClause, it doesn't contain any result operators and is projecting it's MainFromClause only in the selector.
The optimization is the most prominent for queries with Include but will help other cases also.