Fixes #7207 - Query: compilation error for queries with navigations that also compare same navigation to null
Problem was in the optimization that compared navigation to null, specifically:
o.Customer == null => c.CustomerId == null

If we detected that this optimization could be performed, we assumed that the navigation [o.Customer] is no longer needed, and it was being removed from the navigation joins list.
However, if that navigation was also used for another nav access, then it should not have been removed.

Fix is to add counter to the NavigationJoins list, and only remove given navigation join from the list, if it's not referenced anywhere else in the query.

Also removed some dead code and fixed an issue where we would not apply:
c == o.Customer => c.Id == o.CustomerId

if CustomerId was optional (i.e. clr type was different than the PK on the other side)