Fix to #7516 - Query: Compiler crash for invalid Include path
Problem was that when we tried to extract navigation paths specified in the Include we would throw if the Include was on non-entity type.
Fix is to log a warning for this case, just like we do for cases where included entity is not reachable in the final result.
We will still throw exception if user specifies incorrect navigation on existing entity.