Fix to #7131 - Projections With Full Types Not Fixing Up Navigations on Sqlite and SQL Server
Problem was that entities created by MaterializeCollectionNavigation were not being tracked - TrackEntities only works on "naked" entities.

Fix is to add delegate argument to MaterializeCollectionNavigation method that would track the elements of the collection if the query is supposed to be tracked. We already have a similar logic for the Include.