Fix to #8186 - Navs: nav expansion on GJ QSRE in projection doesn't introduce parameter
Problem was that when trying to find query to bind with via outer parameter we were too strict and would not find the appropriate query. We used to share the logic for regular binding and outer parameter binding, however the latter can be more loose and therefore bind to bigger range of queries (e.g. client-side groupjoin).

This also enables additional include scenarios when GJ qsre is projected - fixed a small bug in this area that was exposed by the main fix.