Fix to #8186 - Navs: nav expansion on GJ QSRE in projection doesn't introduce parameter
Problem was that when trying to find query to bind with via outer parameter we were too strict and would not find the appropriate query. However it is also safe to bind to grouping qsre, even though it is wrapped by client GroupJoin.

This also enables additional include scenarios when GJ qsre is projected - fixed a small bug in this area that was exposed by the main fix.