Add ToObservableCollection to LocalView so that it can be used for data binding
Issue #7397

In EF Core, DbSet.Local is a pure view over the state manager. This is much more efficient for normal local queries, etc., but doesn't work for data binding because it does not preserve order or maintain indexes. This change adds DbSet.Local.ToObservableCollection() which creates a less efficient but maintains order and indexes.