Fix to #8754 - Query: manually created GroupJoin with order by key descending generates invalid SQL
Problem was that ordering by the same column twice, but with different direction is invalid. This would usually be user error, however in some cases we introduce order by ourselves (e.g. in case of group join).
If then customer provides their own ordering with opposite direction, we would generate invalid query. We already have logic to catch duplicate ordering (if they are the same).

Fix is to expand the logic that catches "duplicate" ordering to match ordering with different direction and being the same.