Fix to #8369 - Unable to cast object of type 'PropertyExpression' to type 'NullConditionalExpression' for some queries with include, and multiple orderbys
Problem was that we had incorrect assumption in the include compiler when performing orderby lifting. When looking for duplicate orderings we always assumed that subquery projections would be an expression of a particular shape (methood call wrapped around NullConditionalExpression). However, this is not accurate because any expression could be there if it was a previous orderby expression.

Fix is to look for duplicate expression more defensively, using soft rather than explicit cast and also taking MemberExpression into account.