Make HasDefaultValue set that a property is value generated even when passing null or no arguments
Issues #8015, #7872

Setting null or passing no args to the fluent API now sets a sentinel indicating that the method has been called rather than removing the annotation. In most cases the result is functionally equivalent, but in some cases it will now force non-identity value generation without the need to specify what is actually happening in the database.