Add tests that traditional dynamic logging still logs debug messages
It turns out that this still works if and only if the app is not using AddDbContext/UseLoggerFactory. However, if the app is using AddDbContext/UseLoggerFactory and logging is correctly figured in the service provider from which these things come to not filter Debug messages, then this works too. So it's all super weird and hard to understand, but I think we already have the behavior we talked about this morning.

Also, refactored DbContextTest into three files since it was getting too big for R#. Since the diff may be hard to read, the three tests added were:
* Can_log_debug_events_with_OnConfiguring
* Cannot_log_debug_events_with_default_UseLoggerFactory
* Can_log_debug_events_with_UseLoggerFactory_when_configured