Revert setting default log level
Because we now redirect ILoggerFactory in a slightly different way it means that this is not needed when not using an ASP.NET application. When using an ASP.NET application you will get whatever filtering, etc, was setup for ASP.NET. This is different from 1.1 and may be confusing, but doesn't seem unreasonable.