Fix to #9378 - [Test] AsyncQuerySqlCeTest.Where_subquery_on_collection is not robust enough
Improving the test, previously it was doing strict comparison of floating point values which never returned any matches on SQL server, and it accidentally worked because the expected query was doing reference equality of entities, which also never returns anything since the query is not tracked.
The new test performs the same operation as it's sync counterpart.