Allow a global (app-domain) in-memory store
Issue #9613

The internal service provider created by AddDbContext is typically different than the one created when `new` is used. This is because AddDbContext automatically sets ILoggerFactory, etc. from the application service provider, while `new` does not have access to this and so uses a different ILoggerFactory. The same thing can happen with other 'singleton options' like EnableSensitiveDataLogging.

For the in-memory database, this has two consequences:
* Since the store is rooted in the service provider, a new store is created when a situation like above forces a new internal service provider
* Since the store uses `IEntityType` instances to key its table dictionary, these instances must come from the same model to match, which is typically not the case when the model is built twice; once for each service provider

This fix adds a new flag to UseInMemoryDatabase such that the database will really be created globally (i.e. once per app domain) and entity type names are uses as keys instead of IEntityType instances. This could result in memory leaks, but this can be mitigated by seeding with well-known for the test, and/or calling EnsureDeleted appropriately to clean out the data.