Query: Fix for incorrect QueryCache hit when chaining a streaming N+1 query
Issue:
When the query is projecting out a collection object in non-navigation case, we cause N+1 queries.
To support streaming behaviour, we inject QueryableAdapter (or async version) which would be evaluated on when enumerated.

When we are chaining on the navigation after such query, the QueryableAdapter is captured in ConstantExpression.
In our ExpressionEqualityComparer, during GetHashCode, we add such value if ConstantExpression is of type IQueryable. (so that we can fall back to Comparison)
During comparing such ConstantExpressions, we assumed if they are of type IQueryable then they are same. Which is true for IQueryable from DbSet.
But for QueryableAdapter which are also IQueryable, it could still evaluate to different results.
Another thing is, in `EntityQueryableExpressionVisitor` ConstantExpression of EntityQueryable are evaluated. But rest of them are kept intact.
Since we returned equality=true for all IQueryable, every subsequent query even though had different QueryableAdapter which gave different results we made cache hit and re-ran compiled query which has older value of QueryableAdapter (due to EntityQueryableExpressionVisitor) and would evaluate to incorrect values.

Resolves #9301