Infrastructure changes to allow materializer factory factories to use more information
This is a prerequisite to moving type conversions to before the value is put in the buffer rather than when it is read from the buffer. Investigation of supporting more constructor patterns and context injection showed they would likely need these changes too.

Also, ran the SQL Server tests with the typed factory and fixed several issues. This included making exceptions more consistent between typed and untyped factory, and using the property in more places to provide additional info.