Fix #9923 - Perf Regression: AddDbContextPool in asp.net core System.InvalidOperationException
Introduces a "lease" object, which is a scoped object that manages ownership of a pooled context instance. In particular, the lease object is responsible for returning instances to the pool at the end of the scope. This removes the complexity/race condition from DbContext.Dispose and greatly simplifies pooling in general.