Query: Contains should translate to IN only when operand & from expression both are Sql-Compatible
Resolves #10081
Issue: When trying to translate Contains.Item, we convert `x.Blog` to `x.BlogId` since entity comparison is same as identity comparison on entities.
But in case of Contains subquery, we don't visit FromExpression. Instead of validating that FromExpression can be represented in SQL typemapping, we rely on the fact that if translation of Contains.Item can be represented then FromExpression should be.
But that assumption is only valid when the type of Contains.Item does not change during translation. (like scenario above). If type changes (because earlier was not compatible in typemapping), then Contains cannot be translated to InExpression.

This also fixes #9424 in different way since in that case also, Tuple translate to Expression[] changing types.