Query: Translate Distinct after Include to server
Issue: When visiting Selector with _Include method we could not translate it and ended up setting RequiresClientProjection=true which caused Distinct to not get lifted
Fix: _Include should not cause client eval since it would get removed. This allows us to translate Distinct to server.
As a side effect of this in Collection Include case, our distinct order removal logic needs to pattern match more.

Resolves #9912