Query: Apply conditional collapse to comparisons only for truly comparison expressions
Issue:
We have optimization to convert `a ? b : c` where a, b, c are all comparison expression to `a && b || c && d`.
For this case, while b seemed like comparison expression but it was in such pattern due to being nullable bool. (b == true)
A true comparison expression is of type bool since it gives true/false answer. If it is not then its value type expression.
In user case, since we tried incorrect optimization, we ended up with mismatched types in AndAlso expression.

Fix is to improve logic of IsComparisonExpression to only identify boolean expression

Resolves #9468