Query: Fix an issue in SubqueryProjection creation which caused invalid select expression
Issue: All projections in a subquery are required to be ColumnExpression/AliasExpression/ColumnReferenceExpression so that they can be lifted to outer select expression through LiftExpressionFromSubquery method.
In this particular case the expression had convert around it. Because its computed name was unique we did not change it causing invariant to break which later got converted to null projection.

Resolves #10273