Set ProjectStarTable whenever QSRE is being projected out
Resolves #10112

Issue here was, we set ProjectStarTable for QSRE only in subquery. But in this case even though it is top level, it later gets pushed down due to Take-Count sequence causing it to be subquery without projection & projectStarTable set.

Whenever we are projecting QSRE, we can set ProjectStarTable regardless of level. It is just a mapping for us between QSRE returned by queryModel & ProjectStarTable returned by SelectExpression.