Small refactoring to RelationalTypeMapper to handle more metadata
As I have worked on value conversions I have run into the issue that the IProperty is no longer the only input to the type mapping process, since, for example, a value converter may also need to be applied. In the future, things like order preservation are likely to be the same. Therefore, this refactoring introduces parameter object and creates new overloads that using it, while preserving the behavior of existing code that is overriding the old methods.