Query: Fix property binding issue
Issue: When we are binding a property with subSelectExpression, we would just call bind on it and lift the expression but that would be incorrect SQL if the bound expression is not present in projection.
The correct way to bind a property from subSelectExpression is get the expression representing propery in inner projection and lift it.